seperate the parse data components
