+++ website/dev-dist/sw.js
@@ -82,7 +82,7 @@ define(['./workbox-ac6e1177'], (function (workbox) { 'use strict';
     "revision": "3ca0b8505b4bec776b69afdba2768812"
   }, {
     "url": "index.html",
-    "revision": "0.r64drsjvha8"
+    "revision": "0.u9svnhe1dm"
   }], {});
   workbox.cleanupOutdatedCaches();
   workbox.registerRoute(new workbox.NavigationRoute(workbox.createHandlerBoundToURL("index.html"), {
