+++ website/src/components/ParseDataComponents/ParseDataSelectedViewer.jsx
@@ -1,0 +1,55 @@
+import React from "react";
+
+const ParseDataSelectedViewer = ({ selectedFiles, setSelectedFiles }) => {
+  return (
+    <>
+      <div
+        style={{
+          width: "22dvw",
+          height: "100%",
+          backgroundColor: "#929292",
+          border: "1.3dvh solid #2B2B2B",
+          borderRadius: "2.683dvh",
+          overflowY: "auto",
+          overflowX: "hidden",
+        }}
+      >
+        <ul>
+          {selectedFiles.map((file, index) => (
+            <li key={index}>
+              <div
+                style={{
+                  display: "flex",
+                  justifyContent: "space-between",
+                  alignItems: "center",
+                  borderTop: "1.3dvh solid #2B2B2B",
+                }}
+              >
+                <h1 style={{ fontSize: "2.5dvh", marginLeft: "0.64dvw" }}>
+                  {file.name}
+                </h1>
+                <h1
+                  style={{
+                    fontSize: "3.5dvh",
+                    marginRight: "2dvw",
+                    color: "red",
+                    fontWeight: "bold",
+                  }}
+                  onClick={() => {
+                    setSelectedFiles(
+                      selectedFiles.filter((file, i) => i !== index)
+                    );
+                  }}
+                >
+                  x
+                </h1>
+              </div>
+            </li>
+          ))}
+        </ul>
+      </div>
+    </>
+  );
+};
+
+export default ParseDataSelectedViewer;
