--- website/src/test.json
@@ -1,332 +1,0 @@
-[
-  [
-    "alliance",
-    "matchNumber",
-    "scouterInitials",
-    "selectTeam",
-    "noShow",
-    "autoStartPos",
-    "parked",
-    "parkFailed",
-    "climbed",
-    "climbFailed",
-    "trapMadeCount",
-    "trapMissedCount",
-    "comment"
-  ],
-  [
-    "blue",
-    "1",
-    "RIYA K",
-    "2722",
-    false,
-    3,
-    true,
-    false,
-    false,
-    true,
-    0,
-    0,
-    ""
-  ],
-  [
-    "blue",
-    "2",
-    "RIYA K",
-    "2016",
-    false,
-    3,
-    false,
-    false,
-    true,
-    false,
-    0,
-    0,
-    ""
-  ],
-  [
-    "blue",
-    "3",
-    "RIYA K",
-    "2539",
-    false,
-    2,
-    false,
-    false,
-    true,
-    false,
-    1,
-    0,
-    "very good at picking up notes. "
-  ],
-  [
-    "blue",
-    "4",
-    "RIYA K",
-    "5684",
-    false,
-    2,
-    false,
-    true,
-    false,
-    true,
-    0,
-    0,
-    ""
-  ],
-  [
-    "blue",
-    "5",
-    "RIYA K",
-    "6808",
-    false,
-    1,
-    false,
-    true,
-    false,
-    true,
-    0,
-    0,
-    "Had lots of trouble trying to shoot into the speaker, mostly relies on human players to pick up notes"
-  ],
-  [
-    "blue",
-    "7",
-    "RIYA K",
-    "2722",
-    false,
-    1,
-    false,
-    false,
-    true,
-    false,
-    0,
-    0,
-    "Good at scoring in amp and speaker was kind of stuck on the red sie for a while. "
-  ],
-  [
-    "blue",
-    "8",
-    "RIYA K",
-    "5407",
-    false,
-    1,
-    false,
-    false,
-    false,
-    false,
-    0,
-    0,
-    ""
-  ],
-  [
-    "blue",
-    "10",
-    "RIYA K",
-    "708",
-    false,
-    3,
-    false,
-    true,
-    false,
-    true,
-    0,
-    0,
-    "takes a while to load up a note but is able to do far and powerful shots in the speaker. "
-  ],
-  [
-    "blue",
-    "11",
-    "RIYA K",
-    "5684",
-    false,
-    3,
-    false,
-    false,
-    false,
-    false,
-    0,
-    0,
-    "not enough nor accurate data "
-  ],
-  [
-    "blue",
-    "12",
-    "RIYA K",
-    "5401",
-    false,
-    3,
-    true,
-    false,
-    false,
-    true,
-    0,
-    0,
-    "played defense "
-  ],
-  [
-    "blue",
-    "13",
-    "RIYA K",
-    "5407",
-    false,
-    2,
-    false,
-    true,
-    false,
-    true,
-    0,
-    0,
-    ""
-  ],
-  [
-    "blue",
-    "14",
-    "RIYA K",
-    "9094",
-    false,
-    3,
-    false,
-    false,
-    false,
-    false,
-    0,
-    0,
-    ""
-  ],
-  [
-    "blue",
-    "15",
-    "RIYA K",
-    "9015",
-    false,
-    3,
-    true,
-    false,
-    false,
-    true,
-    0,
-    0,
-    ""
-  ],
-  [
-    "blue",
-    "17",
-    "JP",
-    "2590",
-    false,
-    2,
-    false,
-    false,
-    true,
-    false,
-    0,
-    0,
-    "New drivers, robot definitely has potential, uhh drivers do not..."
-  ],
-  [
-    "blue",
-    "18",
-    "JP",
-    "1391",
-    false,
-    4,
-    false,
-    false,
-    true,
-    false,
-    0,
-    0,
-    "No auto, note got stuck inside and the robot couldn't do anything. Attempted defense, but was it against the opposition or their own team!?!?"
-  ],
-  [
-    "blue",
-    "19",
-    "JP",
-    "708",
-    false,
-    3,
-    true,
-    false,
-    false,
-    false,
-    0,
-    0,
-    "Did lots of amp, super powerful shooter, not much to say"
-  ],
-  [
-    "blue",
-    "20",
-    "MT",
-    "193",
-    false,
-    4,
-    false,
-    false,
-    false,
-    false,
-    0,
-    0,
-    "robot was really jittery, something was probably wrong"
-  ],
-  [
-    "red",
-    "24",
-    "JP",
-    "8714",
-    false,
-    4,
-    false,
-    false,
-    true,
-    false,
-    0,
-    0,
-    "Note got stuck in intake. Definitely some new drivers operating the robot. Attempted defense, I'd give it a 4/10. The climb was very weird, but it worked I guess."
-  ],
-  [
-    "red",
-    "25",
-    "JP",
-    "293",
-    false,
-    1,
-    true,
-    false,
-    false,
-    false,
-    0,
-    0,
-    ""
-  ],
-  [
-    "red",
-    "26",
-    "JP",
-    "6808",
-    false,
-    3,
-    true,
-    false,
-    false,
-    false,
-    0,
-    0,
-    "One word. Sad."
-  ],
-  [
-    "red",
-    "27",
-    "JP",
-    "365",
-    false,
-    2,
-    false,
-    false,
-    true,
-    false,
-    0,
-    0,
-    "Weird launching, not a lot of game sense. Drivers are kinda mid."
-  ]
-]
